Корректировка страниц

+++ actions/admin/del_user.php
@@ -1,12 +1,19 @@
 <?php
 require_once '../../config.php';
 
+$currentId = $_SESSION['user']['id'];
 $userId = $_POST['id'];
 
-$query = "DELETE FROM users WHERE id = :id";
-$res = $pdo->prepare($query);
-$res->execute([
+if ($userId != $currentId) {
+    $query = "DELETE FROM users WHERE id = :id";
+    $res = $pdo->prepare($query);
+    $res->execute([
     ':id' => $userId
     ]);
-header('location: /pages/admin/users.php');
+    header('location: /pages/admin/users.php');
+} else {
+    $_SESSION['delError'] = true;
+    header('location: /pages/admin/users.php');
+}
+
